Reflash automatically after 12 seconds
Instantiates the USB drive on your host PC,
so that a new version of the .UF2 can be
dragged and dropped.

Uses Gammon's idea for an empty .ino sketch,
with the real sketch in a companion .cpp file.

Used 'ag' to find the right code base, which
seemed to be:

 $ ag usb_boot | cat -n
     1  cores/rp2040/SerialUSB.cpp:294:        reset_usb_boot(0,0);

Then just swiped all the stuff from SerialUSB.cpp and put
it into the sketch, without examination of any sort.

'Just worked'.  Probably quite 'wrong' of course.  Quick test
of the concept.

	modified:   _test-a.d/fade-test-ac/fade-test-ac.ino
	new file:   _test-a.d/fade-test-ac/fade-test-ac_SKETCH.cpp

On branch dvlp-aa