Extensive EXCEPTION handling
Flood the RAM with instructions that handle
boundary jumps.  If the instruction pointer
lands still within the RAM, but past the
program, should generally catch the overjump
and handle an exception.

Regular program contents overwrites the flood,
so that there's no interference.

Like a canvas base layer; 'paint' an exception
handling 'background' as it were. ;)

	modified:   n_monitor-ff_SKETCH.cpp

On branch dvlp-aa