Nice.  Shows how allocated ram array RAMSPACE[] is populated from both the contents of memory[] and some pre-writes 'underneath' just prior.
Since 'memory[]' is the last thing to be copied
into 'RAMSPACE[]' it prevails.  However, that
copy does not overwrite stuff outside the
 .. dimensions .. of 'memory[]' as it is mapped
into 'RAMSPACE[]' (not so much 'mapped' as copied).

Since other stuff is written within the bounds of
RAMSPACE[] that stuff remains as-written.

0x4F4F4F (FILL_CHAR), 0x646E45EE (op_hxee)
and 0x706F6E (op_nop) are all included in this
'other stuff'.

At times, 0x2222 and 0xced1 (op_exc) can be found.

This 'other stuff' .. is found at the very end
of RAMSPACE[] .. start at its end, and work backwards
to find the boundary of the image of the contents
of 'memory[]' that were copied into 'RAMSPACE[]'.

	modified:   n_monitor-ff_SKETCH.cpp

On branch dvlp-aa