program_boundary fence set low at 180
Have not investigated where it should be
set for maximum use of the allocated space.

It'll probably be related closely to RAM_SIZE
further along in development.

hex 21BC decimal 180 4 * hex + . 248C  ok

That makes it 180 array elements (the subscript
to the array in the square braces) i.e. 4 x 180
is 720 bytes for this program, from addresses
200021BC through 2000248C (with a few spares after
the active code region).

Last active code line in the dump is:

2000246C: 6C 69 74  3  7D 12 00 00  64 6C 79  3  72 66 6C  3  lit.}...dly.rfl.

the op_rfl instruction (reflash) ends the program. ;)

	modified:   n_monitor-gg_SKETCH.cpp

On branch dvlp-aa